solarmaker/app/requirements.txt
Django==3.2.6
gunicorn==20.1.0
psycopg2-binary==2.9.1
djangorestframework==3.13.1
